Reduce worker event flush overhead
